bug-hunt #76: -Wcast-qual — fix remaining qualifier discarding (round 2)
Fixes 72 more -Wcast-qual violations across 27 additional files:
- string_ext.c: strcasestr, strchrnul, memmem
- stdlib_ext.c: strtoll, strtoull
- search.c: bsearch, lfind, lsearch
- errno_ext.c: strerror
- pthread.c: futex calls & rwlock patterns
- pstore.c, kdump.c, kexec.c: volatile-memset patterns
- firmware.c, keyring.c, sort_ext.c
- refcount_ext.h, bitops.h: inline header fixes
- io_map.c, fault_inject.c, kunit.h, kunit_tests.c
- usb_hid.c, usb_cdc_acm.c, usb_uas.c
- efi_runtime.c, uefi_runtime.c
- fq_codel.c: const packet enqueue
- dm.c: const char** cast

Build verified clean with -Wcast-qual -Werror across entire kernel.

+++ src/drivers/usb_hid.c
@@ -288,7 +288,7 @@ int usb_hid_set_report(uint8_t report_type, uint8_t report_id,
     uint8_t iface = g_keyboard_present ? g_kbd_intf : g_mouse_intf;
 
     return usb_control(g_op_base, HID_REQTYPE_SET, HID_REQ_SET_REPORT,
-                        w_val, iface, (uint16_t)len, (void *)buf, g_dev_addr);
+                        w_val, iface, (uint16_t)len, (void *)(uintptr_t)buf, g_dev_addr);
 }
 
 /*
